commands refactor fix

+++ cloud-client/src/main/java/my/cloud/client/service/impl/commands/FilesOfferRefusedCommand.java
@@ -1,27 +1,43 @@
 package my.cloud.client.service.impl.commands;
 
 import command.domain.CommandCode;
+import command.service.CommandService;
 import files.domain.TransferId;
+import io.netty.channel.ChannelHandlerContext;
 import my.cloud.client.factory.Factory;
-import my.cloud.client.service.impl.commands.base.BaseClientCommand;
+
+import java.util.function.Consumer;
+
+import static my.cloud.client.service.impl.commands.util.ClientCommandUtil.*;
 
 /**
  * Called when server cant accept upload
  */
-public class FilesOfferRefusedCommand extends BaseClientCommand {
+public class FilesOfferRefusedCommand implements CommandService {
 
-    public FilesOfferRefusedCommand() {
-        expectedArgumentsCountCheck = i -> i > 0;
-    }
+    private Consumer<String[]> consumer;
 
     @Override
-    protected void processArguments(String[] args) {
+    public void processCommand(ChannelHandlerContext ctx, String[] args) {
+        if (wrongArgumentsLength(args, i -> i == 0)) {
+            return;
+        }
+
         TransferId t = Factory.getFileTransferAuthService().getTransferIfValid(args[0]);
         Factory.getUploadProgressService().remove(t.origin);
 
         for (int i = 1; i < args.length; i++) {
             Factory.getFileTransferAuthService().getTransferIfValid(args[i]);
         }
+
+        if (consumer != null) {
+            consumer.accept(args);
+        }
+    }
+
+    @Override
+    public void setListener(Consumer<String[]> consumer) {
+        this.consumer = consumer;
     }
 
     @Override
